+++ src/Symfony/Component/Validator/Constraints/TypeValidator.php
@@ -33,22 +33,25 @@ class TypeValidator extends ConstraintValidator
             return;
         }
 
-        $type = strtolower($constraint->type);
-        $type = 'boolean' == $type ? 'bool' : $constraint->type;
-        $isFunction = 'is_'.$type;
-        $ctypeFunction = 'ctype_'.$type;
+        $types = (array) $constraint->type;
 
-        if (\function_exists($isFunction) && $isFunction($value)) {
-            return;
-        } elseif (\function_exists($ctypeFunction) && $ctypeFunction($value)) {
-            return;
-        } elseif ($value instanceof $constraint->type) {
-            return;
+        foreach ($types as $type) {
+            $type = strtolower($type);
+            $type = 'boolean' === $type ? 'bool' : $type;
+            $isFunction = 'is_'.$type;
+            $ctypeFunction = 'ctype_'.$type;
+            if (\function_exists($isFunction) && $isFunction($value)) {
+                return;
+            } elseif (\function_exists($ctypeFunction) && $ctypeFunction($value)) {
+                return;
+            } elseif ($value instanceof $type) {
+                return;
+            }
         }
 
         $this->context->buildViolation($constraint->message)
             ->setParameter('{{ value }}', $this->formatValue($value))
-            ->setParameter('{{ type }}', $constraint->type)
+            ->setParameter('{{ type }}', implode('|', $types))
             ->setCode(Type::INVALID_TYPE_ERROR)
             ->addViolation();
     }