+++ src/Symfony/Component/Validator/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * added the `compared_value_path` parameter in violations when using any 
    comparison constraint with the `propertyPath` option.
+ * added support for checking an array of types in `TypeValidator`
 
 4.3.0
 -----